+++ src/components/Progress/mixins/progressBarProps.ts
@@ -1,43 +1,38 @@
 import type { Prop } from 'vue';
-import { booleanProp, booleanTrueProp, stringOrNumberProp, stringProp } from '../../../mixins/CommonProps';
-import type { TProgressBarValuePosition } from '../types';
+import { booleanProp, stringOrNumberProp, stringProp } from '../../../mixins/CommonProps';
+import type {
+    TProgressBarLabelPosition,
+    TProgressBarValuePosition,
+    TTextLabelAlignment,
+} from '../types';
 
 export const progressBarProps = {
-    /**
-     * The component color appearance.
-     */
     color: stringProp,
-    /**
-     * The ProgressBar thickness.
-     */
     height: stringOrNumberProp,
-    /**
-     * The value monitored by `v-model` to control the progress bar value.
-     */
+    innerCls: stringProp,
+    label: stringProp,
+    labelAlignment: {
+        type: String,
+        default: 'center',
+        validator: (value: string) => ['start', 'end', 'center'].includes(value),
+    } as Prop<TTextLabelAlignment>,
+    labelPosition: {
+        type: String,
+        default: 'top',
+        validator: (value: string) => ['start', 'end', 'top', 'bottom'].includes(value),
+    } as Prop<TProgressBarLabelPosition>,
+    roundedOff: booleanProp,
+    striped: booleanProp,
+    stripedAnimation: booleanProp,
+    showValue: booleanProp,
     modelValue: {
         type: Number,
         default: 0,
-        validator: (value: number): boolean => value >= 0 && value <= 100
+        validator: (value: number): boolean => value >= 0 && value <= 100,
     },
-    /**
-     * Set to `false` to remove the rounded border on the side of the progress bar.
-     */
-    rounded: booleanTrueProp,
-    /**
-     * Create striped ProgressBar.
-     */
-    striped: booleanProp,
-    /**
-     * Create animated stripe ProgressBar.
-     */
-    stripedAnimation: booleanProp,
-    /**
-     * Display progress bar's value or not.
-     */
-    showValue: booleanProp,
     valuePosition: {
         type: String,
         default: 'inside',
-        validator: (value: string) => ['start', 'end', 'top', 'bottom', 'inside'].includes(value)
-    } as Prop<TProgressBarValuePosition>
-}
+        validator: (value: string) => ['start', 'end', 'top', 'bottom', 'inside'].includes(value),
+    } as Prop<TProgressBarValuePosition>,
+};